Move ICompletionRules to Microsoft.CodeAnalysis.Completion namespace
