+++ src/EditorFeatures/Core/Implementation/IntelliSense/Completion/Controller.Session.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using Microsoft.CodeAnalysis.Completion.Rules;
+using Microsoft.CodeAnalysis.Completion;
 using Microsoft.VisualStudio.Text;
 using Roslyn.Utilities;
