+++ src/Features/Core/Completion/ICompletionRules.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-namespace Microsoft.CodeAnalysis.Completion.Rules
+namespace Microsoft.CodeAnalysis.Completion
 {
     internal interface ICompletionRules
     {
